requirements.txt
playwright==1.44.0
openai==1.35.0
python-dotenv==1.0.0
requests>=2.31.0
